First commit
This is the implementation of a typed prolog. The version I committed
here works, and I'm implementing a way of parametrizing the constraint
in the definition of a term. That's because I was using Show (Term a) as
constraint, and later I discovered that I also need Unifiable (Term a)
to write the unification algorithm, and so I'd like to parametrize the
choice.