Refactor substitutions to depend on WellFormedness
WellFormed a is a typeclass introduced to imply `Show (Term a)`, which I
need for the Show instances, and `Unifiable (Term a)`, which I need for
the unification machinery.

The technical point to remember is that superclass implication is only
read by superclasses in class definition, not in instances, which only
expose the superclass into scope in their where clause.