change eval/run unification to be consistent
with the names in Control.Monad.State. There is now a top-level unify
function that does the unification of two terms. This is because, as we
don't do the occur check for performance, the substitutions may contain
variable cycles, and so it's important that we begin with an empty
substitution every time.