complete Substitutable typeclass refactoring
using defaultInstances now we can hide the function sbs without
generating errors. Beside that, we have now `Substitable a` instead of
`Substitutable (Term a)`, which limits us in some way (because we can't
say `Substitutable (Term' a)` where Term' has say attributed variables),
but on the other hand it makes signatures simpler, and it might be
easier to understand.